+++ packages/vuikit/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## WIP
+
+ - Fixed `vk-tooltip` unbind.
+
 ## 0.8.4 (April 26, 2018)
 
  - Fixed regression from 0.8.3